+++ Tests/MapboxMapsTests/MapView/Integration Tests/Map/MapViewIntegrationTests.swift
@@ -30,6 +30,10 @@ class MapViewIntegrationTests: IntegrationTestCase {
         rootView = nil
 
         try super.tearDownWithError()
+
+        if let url = dataPathURL {
+            removeFilesInDirectoryTree(at: url)
+        }
     }
 
     func testMapViewIsReleasedAfterCameraTransition() throws {
