+++ Tests/MapboxMapsTests/MapView/Integration Tests/Map/MapboxMapIntegrationTests.swift
@@ -20,6 +20,14 @@ class MapboxMapIntegrationTests: IntegrationTestCase {
         rootView = root
     }
 
+    override func tearDownWithError() throws {
+        try super.tearDownWithError()
+
+        if let url = dataPathURL {
+            removeFilesInDirectoryTree(at: url)
+        }
+    }
+
     // MARK: - Tests
 
     func testLoadStyleURICompletionIsCalled() {
