+++ Tests/MapboxMapsTests/Offline/TileStore+Tests.swift
@@ -18,30 +18,31 @@ extension XCTestCase {
 
         return cacheDirectoryURL
     }
-}
 
-//internal func removeDirectory(at fileURL: URL) throws {
-//    let fileManager = FileManager.default
-//    let resourceKeys = Set<URLResourceKey>([.isDirectoryKey])
-//    var fileURLs: [URL] = []
-//
-//    guard let enumerator = fileManager.enumerator(at: fileURL, includingPropertiesForKeys: Array(resourceKeys)) else {
-//        return
-//    }
-//
-//    for case let pathURL as URL in enumerator {
-//        let resourceValues = try? pathURL.resourceValues(forKeys: resourceKeys)
-//
-//        if let isDirectory = resourceValues?.isDirectory, isDirectory {
-//            continue
-//        }
-//
-//        fileURLs.append(pathURL)
-//    }
-//
-//    // Remove files
-//    try fileURLs.forEach {
-//        try fileManager.removeItem(at: $0)
-//    }
-//    try fileManager.removeItem(at: fileURL)
-//}
+    internal func removeFilesInDirectoryTree(at fileURL: URL) throws {
+        let fileManager = FileManager.default
+        let resourceKeys = Set<URLResourceKey>([.isDirectoryKey])
+        var fileURLs: [URL] = []
+
+        guard let enumerator = fileManager.enumerator(at: fileURL, includingPropertiesForKeys: Array(resourceKeys)) else {
+            return
+        }
+
+        for case let pathURL as URL in enumerator {
+            let resourceValues = try? pathURL.resourceValues(forKeys: resourceKeys)
+
+            if let isDirectory = resourceValues?.isDirectory, isDirectory {
+                continue
+            }
+
+            fileURLs.append(pathURL)
+        }
+
+        // Remove files
+        try fileURLs.forEach {
+            try fileManager.removeItem(at: $0)
+        }
+
+        //try fileManager.removeItem(at: fileURL)
+    }
+}
